Split out Response with trait
